refactor: move SimulationConfig to a new file and update import paths
